# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assignments/{assignmentId}/practice:
    post:
      operationId: createPracticeExam
      tags: [practice]
      summary: Generate a fresh, infinite practice version of a graded assignment
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeExam"
...
  /practice-exams/{sessionId}:
    get:
      operationId: getPracticeExam
      tags: [practice]
      summary: Get a practice exam session with its problems
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeExam"
  /practice-exams/{sessionId}/submit:
    post:
      operationId: submitPracticeExam
      tags: [practice]
      summary: Submit a practice exam; returns heavy per-problem + overall feedback and focus pointers
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PracticeExamSubmitInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeExamResult"
  /practice-exams/{sessionId}/feedback:
    get:
      operationId: getFeedbackMessages
      tags: [practice]
      summary: Get the feedback dialogue history for a practice exam
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeedbackMessage"
  /practice-exams/{sessionId}/feedback/ask:
    post:
      operationId: askFeedback
      tags: [practice]
      summary: Ask a follow-up question about the feedback; returns the coach's reply
      parameters:
        - name: sessionId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackAskInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackMessage"
...
components:
  schemas:
...
    KeystrokeTrace:
      type: object
      required: [keystrokeCount, eraseCount, durationMs]
      properties:
        keystrokeCount: { type: integer }
        eraseCount: { type: integer }
        bulkInsertCount: { type: integer, default: 0 }
        longestBulkInsertChars: { type: integer, default: 0 }
        rewriteSegments: { type: integer, default: 0 }
        durationMs: { type: integer }
...
    PracticeExam:
      type: object
      required: [id, assignmentId, assignmentTitle, kind, weekNumber, status, createdAt, problems]
      properties:
        id: { type: integer }
        assignmentId: { type: integer }
        assignmentTitle: { type: string }
        kind:
          type: string
          enum: [homework, test, midterm, final]
        weekNumber: { type: integer }
        status:
          type: string
          enum: [in_progress, submitted]
        createdAt: { type: string, format: date-time }
        problems:
          type: array
          items:
            $ref: "#/components/schemas/PracticeExamProblem"
    PracticeExamProblem:
      type: object
      required: [id, position, prompt, topicId, difficulty]
      properties:
        id: { type: integer }
        position: { type: integer }
        prompt: { type: string }
        topicId: { type: integer }
        topicTitle: { type: ["string", "null"] }
        difficulty: { type: number }
    PracticeExamSubmitInput:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/PracticeExamAnswerInput"
    PracticeExamAnswerInput:
      type: object
      required: [problemId, answer, trace]
      properties:
        problemId: { type: integer }
        answer: { type: string }
        trace:
          $ref: "#/components/schemas/KeystrokeTrace"
    PracticeExamResult:
      type: object
      required: [sessionId, score, total, percent, perProblem, overallFeedback, focusPointers, encouragement]
      properties:
        sessionId: { type: integer }
        score: { type: number }
        total: { type: number }
        percent: { type: number }
        perProblem:
          type: array
          items:
            $ref: "#/components/schemas/PracticeExamProblemFeedback"
        overallFeedback: { type: string }
        focusPointers:
          type: array
          items:
            $ref: "#/components/schemas/FocusPointer"
        encouragement: { type: string }
    PracticeExamProblemFeedback:
      type: object
      required: [problemId, prompt, correct, feedback, correctAnswer, explanation]
      properties:
        problemId: { type: integer }
        prompt: { type: string }
        topicTitle: { type: ["string", "null"] }
        userAnswer: { type: string }
        correctAnswer: { type: string }
        correct: { type: boolean }
        feedback: { type: string }
        explanation: { type: string }
    FocusPointer:
      type: object
      required: [topicTitle, severity, why, action]
      properties:
        topicId: { type: ["integer", "null"] }
        topicTitle: { type: string }
        severity:
          type: string
          enum: [critical, important, minor]
        why: { type: string }
        action: { type: string }
    FeedbackMessage:
      type: object
      required: [id, role, content, at]
      properties:
        id: { type: integer }
        role:
          type: string
          enum: [student, coach]
        content: { type: string }
        at: { type: string, format: date-time }
    FeedbackAskInput:
      type: object
      required: [message]
      properties:
        message: { type: string }
...