# lib/api-zod/src/generated/types/feedbackMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackMessageRole } from './feedbackMessageRole';

export interface FeedbackMessage {
  id: number;
  role: FeedbackMessageRole;
  content: string;
  at: Date;
}