# lib/api-zod/src/generated/types/practiceExam.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PracticeExamKind } from './practiceExamKind';
import type { PracticeExamProblem } from './practiceExamProblem';
import type { PracticeExamStatus } from './practiceExamStatus';

export interface PracticeExam {
  id: number;
  assignmentId: number;
  assignmentTitle: string;
  kind: PracticeExamKind;
  weekNumber: number;
  status: PracticeExamStatus;
  createdAt: Date;
  problems: PracticeExamProblem[];
}