# lib/api-zod/src/generated/types/practiceExamAnswerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { KeystrokeTrace } from './keystrokeTrace';

export interface PracticeExamAnswerInput {
  problemId: number;
  answer: string;
  trace: KeystrokeTrace;
}