# lib/api-zod/src/generated/types/practiceExamProblem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PracticeExamProblem {
  id: number;
  position: number;
  prompt: string;
  topicId: number;
  /** @nullable */
  topicTitle?: string | null;
  difficulty: number;
}