# lib/api-zod/src/generated/types/practiceExamProblemFeedback.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PracticeExamProblemFeedback {
  problemId: number;
  prompt: string;
  /** @nullable */
  topicTitle?: string | null;
  userAnswer?: string;
  correctAnswer: string;
  correct: boolean;
  feedback: string;
  explanation: string;
}