# lib/api-zod/src/generated/types/practiceExamResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FocusPointer } from './focusPointer';
import type { PracticeExamProblemFeedback } from './practiceExamProblemFeedback';

export interface PracticeExamResult {
  sessionId: number;
  score: number;
  total: number;
  percent: number;
  perProblem: PracticeExamProblemFeedback[];
  overallFeedback: string;
  focusPointers: FocusPointer[];
  encouragement: string;
}